Rename arwendebug to vmserver
