+++ .gitignore
@@ -17,9 +17,9 @@
 /cmd/test/test
 /cmd/testgen/dns/dns
 /cmd/arwen/arwen
-/arwendebug/testdata/db
-/cmd/arwendebug/arwendebug
-/cmd/arwendebug/db
+/vmserver/testdata/db
+/cmd/vmserver/vmserver
+/cmd/vmserver/db
 /ipc/tests/testdata/streams
 /ipc/tests/arwen
