+++ vmserver/common_test.go
@@ -1,4 +1,4 @@
-package arwendebug
+package vmserver
 
 import (
 	"testing"
