+++ vmserver/constants.go
@@ -1,4 +1,4 @@
-package arwendebug
+package vmserver
 
 // DefaultGasPrice is the default gas price for debugging
 const DefaultGasPrice = 200000000000
