+++ vmserver/database.go
@@ -1,4 +1,4 @@
-package arwendebug
+package vmserver
 
 import (
 	"encoding/json"
