+++ vmserver/messagesBase.go
@@ -1,4 +1,4 @@
-package arwendebug
+package vmserver
 
 import (
 	"math/big"
