+++ vmserver/server.go
@@ -1,4 +1,4 @@
-package arwendebug
+package vmserver
 
 import (
 	"fmt"
